separate dashboard components
