+++ src/components/Dashboard.js
@@ -3,10 +3,10 @@ import { VictoryLine, VictoryChart, VictoryTheme } from 'victory';
 import Box from './Box'
 import DatePicker from "react-datepicker";
 import "react-datepicker/dist/react-datepicker.css";
-import ByIntent from './ByIntent'
-import donationsCircle from "../images/donations-circle.svg";
-import twoHandsDonation from "../images/two-hands-donation.svg";
-import Steps from "../images/steps.png";
+import BySource from './BySource'
+import DonationAmount from "./DonationAmount"
+import KeyStatistics from "./KeyStatistics" 
+import ByIntent from "./ByIntent"
 
 const fakeData = [
     { x: "cats", y: 55, fill: "#0a2" },
@@ -36,64 +36,15 @@ const fakeData = [
             </div>
         </div>
 
-        <div className = "wrap">     
-        <h1>Donation Amount</h1>
-          <Box>
-          <div className="line-chart">
-          
-              <VictoryChart
-              theme={VictoryTheme.material}
-              height={200}
-              width={300}
-              >
-              <VictoryLine
-                  style={{
-                  data: { stroke: "#FE5366" },
-                  parent: { border: "1px solid #CC7503"},
-                  }}
-                  data={[
-                  { x: 1, y: 2 },
-                  { x: 2, y: 3 },
-                  { x: 3, y: 5 },
-                  { x: 4, y: 4 },
-                  { x: 5, y: 7 }
-                  ]}
-              />
-              </VictoryChart>
-          </div>
-          </Box>
-
-          <div className = "keystats">
-          <h1>Key Statistics</h1>
-            <Box>
-            <div>
-               <img src={donationsCircle} className="donationscircle" />
-               <div className = "totaldonationamt">$12,154</div>
-                 <div className = "keystatslabel">Total Donation Amount</div>
-            </div>
-
-            </Box>
-
-            <Box>
-            <div>
-               <img src={twoHandsDonation} className="twoHandsDonation" /></div>
-               <div className = "totalnumberdonors">6329</div>
-                <div className = "keystatslabel">Total Number of Donors</div>
+        <div className = "wrap"> 
             
+          <DonationAmount />
 
-            </Box>
-            
-            <img src={Steps} className="steps" />
-
-          </div>
-
-          <h2>Total Number of Donations</h2>
-          <h1>By Intent</h1>
-            <Box>
-              <ByIntent />
-            </Box>
-
+          <KeyStatistics />
           
+          <BySource />    
+
+          <ByIntent />    
           
         </div>
       </>
